sửa kiểu dữ liệu cho biến 'result'

+++ B. Thực hiện từng đề bài riêng/main.cpp
@@ -109,7 +109,7 @@ void releaseMemory(double *vectorA, double *vectorB)
 }
 double sequentialDotProduct(double *vectorA, double *vectorB, int size)
 {
-    double result = 0.0;
+    long double result = 0.0;
 
     for (int i = 0; i < size; i++)
     {
